+++ src/main/kotlin/icu/windea/pls/lang/inspections/localisation/common/UnsupportedRecursionInspection.kt
@@ -1,4 +1,4 @@
-package icu.windea.pls.lang.inspections.localisation.bug
+package icu.windea.pls.lang.inspections.localisation.common
 
 import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.ProblemsHolder
@@ -14,6 +14,7 @@ import icu.windea.pls.localisation.psi.ParadoxLocalisationProperty
 
 /**
  * （对于本地化文件）检查是否存在不支持的递归。
+ *
  * - 对于每个本地化，检查其本地化文本中是否存在递归的本地化引用。
  */
 class UnsupportedRecursionInspection : LocalInspectionTool(), DumbAware {