+++ src/main/kotlin/icu/windea/pls/lang/inspections/script/inlineScript/ConflictingInlineScriptUsageInspection.kt
@@ -1,26 +1,19 @@
-package icu.windea.pls.lang.inspections.script.inference
+package icu.windea.pls.lang.inspections.script.inlineScript
 
 import com.intellij.codeInspection.InspectionManager
-import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiFile
 import icu.windea.pls.PlsBundle
 import icu.windea.pls.config.configContext.inlineScriptHasConflict
 import icu.windea.pls.lang.quickfix.GotoInlineScriptUsagesFix
-import icu.windea.pls.lang.selectRootFile
 import icu.windea.pls.lang.util.ParadoxExpressionManager
 import icu.windea.pls.lang.util.ParadoxInlineScriptManager
 
 /**
- * 检查内联脚本的使用位置是否存在冲突。
+ * 检查内联脚本用法是否存在冲突。
  */
-class ConflictingInlineScriptUsageInspection : LocalInspectionTool() {
-    override fun isAvailableForFile(file: PsiFile): Boolean {
-        if (selectRootFile(file) == null) return false
-        return true
-    }
-
+class ConflictingInlineScriptUsageInspection : InlineScriptInspectionBase() {
     override fun checkFile(file: PsiFile, manager: InspectionManager, isOnTheFly: Boolean): Array<ProblemDescriptor>? {
         val inlineScriptExpression = ParadoxInlineScriptManager.getInlineScriptExpression(file) ?: return null
         val configContext = ParadoxExpressionManager.getConfigContext(file) ?: return null