+++ src/main/kotlin/icu/windea/pls/lang/inspections/script/inlineScript/UnsupportedInlineScriptUsageInspection.kt
@@ -1,21 +1,22 @@
-package icu.windea.pls.lang.inspections.script.bug
+package icu.windea.pls.lang.inspections.script.inlineScript
 
-import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.ProblemsHolder
-import com.intellij.openapi.project.DumbAware
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiElementVisitor
 import com.intellij.psi.PsiFile
 import icu.windea.pls.PlsBundle
+import icu.windea.pls.lang.selectGameType
 import icu.windea.pls.lang.selectRootFile
-import icu.windea.pls.script.psi.ParadoxScriptInlineMath
+import icu.windea.pls.lang.util.ParadoxInlineScriptManager
+import icu.windea.pls.script.psi.ParadoxScriptProperty
 
 /**
- * （对于脚本文件）检查是否在不支持的地方使用了内联数学表达式。
+ * 检查是否在不支持的地方使用了内联脚本。
  */
-class UnsupportedInlineMathInspection : LocalInspectionTool(), DumbAware {
+class UnsupportedInlineScriptUsageInspection : InlineScriptInspectionBase()/*, DumbAware*/ {
     override fun isAvailableForFile(file: PsiFile): Boolean {
         if (selectRootFile(file) == null) return false
+        if (!ParadoxInlineScriptManager.isSupported(selectGameType(file))) return false
         return true
     }
 
@@ -24,13 +25,12 @@ class UnsupportedInlineMathInspection : LocalInspectionTool(), DumbAware {
         if (extension == "asset") {
             return object : PsiElementVisitor() {
                 override fun visitElement(element: PsiElement) {
-                    if (element is ParadoxScriptInlineMath) {
-                        holder.registerProblem(element, PlsBundle.message("inspection.script.unsupportedInlineMath.desc.1"))
-                    }
+                    if (element !is ParadoxScriptProperty) return
+                    if (!ParadoxInlineScriptManager.isMatched(element.name)) return
+                    holder.registerProblem(element, PlsBundle.message("inspection.script.unsupportedInlineScriptUsage.desc.1"))
                 }
             }
         }
-
         return PsiElementVisitor.EMPTY_VISITOR
     }
 }