+++ CHANGELOG.md
@@ -301,7 +301,7 @@
 
 - [X] 兼容脚本文件中的内联模板表达式（如，`has_ethic = ethic_[[fanatic]fanatic_]pacifist`）
 - [X] #96
-- [X] #97 兼容VIC3中的定义（define）引用（对应新的规则类型`$define_reference`）
+- [X] #97 兼容VIC3中的预设值（define）引用（对应新的规则类型`$define_reference`）
 - [X] #100 删除`disableLogger`以免忽略某些重要的报错
 
 ***